moved flat_datasets to catalog
